incorporated context api, added sun icon to darkmode toggle, added sun icon and darkmode toggle to navbar
