+++ src/sass/_nav.scss
@@ -3,16 +3,10 @@ nav {
   justify-content: space-evenly;
   align-items: center;
   height: 8vh;
-
   // NOTE Light Mode Theme borders
   border-bottom: 3px solid $base-color;
   border-top: 10px solid $base-color;
   background-color: white;
-
-  // NOTE Dark Mode Theme borders
-  // border-bottom: 0px solid $base-color;
-  // border-top: 10px solid $base-color;
-  // background-color: #222;
 
   border-bottom-left-radius: 20px;
   border-bottom-right-radius: 20px;
@@ -42,9 +36,18 @@ nav {
   }
 }
 
+.mode-change-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  .fas.fa-sun {
+    color: orange;
+  }
+}
+
 .dark-mode-nav {
   // NOTE Dark Mode Theme borders
-  border-bottom: 0px solid $base-color;
+  border-bottom: 5px solid $base-color;
   border-top: 10px solid $base-color;
   background-color: #222;
 }
