+++ src/sass/_welcome-page.scss
@@ -217,3 +217,19 @@
   font-size: 1.5rem;
   text-align: center;
 }
+
+.welcome-page-mode-toggle {
+  display: flex;
+  position: relative;
+}
+
+.welcome-page-mode-toggle {
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  color: orange;
+  font-size: 1.2rem;
+  padding-top: 1%;
+  padding-left: 1%;
+}
